# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Polemicyst API
  description: API for the Polemicyst viral clip generation platform
  version: 1.0.0
servers:
  - url: http://10.0.2.2:3000
    description: Android emulator (local dev)
  - url: https://polemicyst.com
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TriggerClipRequest:
      type: object
      required: [feedVideoId, userId]
      properties:
        feedVideoId:
          type: string
        userId:
          type: string
        aspectRatio:
          type: string
          nullable: true
        scoringMode:
          type: string
          enum: [heuristic, hybrid, gemini]
        includeAudio:
          type: boolean
        saferClips:
          type: boolean
        targetPlatform:
          type: string
          enum: [all, reels, shorts, youtube]
        contentStyle:
          type: string
          enum: [auto, politics, comedy, education, podcast, gaming, vlog, other]
        llmProvider:
          type: string
          enum: [gemini, ollama]
        clipLength:
          type: string
        minCandidates:
          type: integer
        maxCandidates:
          type: integer
        minScore:
          type: number
        percentile:
          type: number
        maxGeminiCandidates:
          type: integer
    TriggerClipResponse:
      type: object
      required: [message, jobId]
      properties:
        message:
          type: string
        jobId:
          type: string
...
paths:
...
  /api/feedVideos/{id}:
    delete:
      operationId: deleteFeedVideo
      summary: Delete a feed video
      tags: [FeedVideos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /api/trigger-clip:
    post:
      operationId: triggerClipGeneration
      summary: Enqueue a clip generation job
      tags: [Clips]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerClipRequest'
      responses:
        '200':
          description: Job enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerClipResponse'
...
  /api/admin/logs:
    get:
      operationId: getAdminLogs
      summary: Get job logs (admin only)
      tags: [Admin]
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
        - name: jobType
          in: query
          schema:
            type: string
            enum: [transcription, speaker-transcription, clip-generation, download]
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, started, completed, failed]
        - name: feedVideoId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Job logs with summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        feedVideoId:
                          type: string
                        jobType:
                          type: string
                        status:
                          type: string
                        message:
                          type: string
                          nullable: true
                        error:
                          type: string
                          nullable: true
                        durationMs:
                          type: integer
                          nullable: true
                        createdAt:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  summary:
                    type: array
                    items:
                      type: object
                      properties:
                        jobType:
                          type: string
                        status:
                          type: string
                        count:
                          type: integer
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...